requirements.txt
fastapi
uvicorn
pydantic-ai[google]
python-dotenv
jinja2